Refactor: useCountdown로 리팩토링
