+++ src/components/WindowSizeExample.jsx
@@ -2,7 +2,7 @@ import { useWindowSize } from "../hooks/useWindowSize";
 
 export const WindowSizeExample = () => {
   // 실습 1. 하단 코드를 useWindowSize (커스텀 훅으로 바꿔주세요!)
-  const  windowSize  = useWindowSize();
+  const {windowSize} = useWindowSize();
   return (
     <div>
       <h2>useWindowSize 실습</h2>
