+++ src/hooks/useWindowSize.js
@@ -23,5 +23,5 @@ export const useWindowSize = () => {
     };
   }, []);
 
-  return  windowSize ;
+  return  {windowSize} ;
 };
